FMOD setup with first music objects (#65 #90)
FMOD setup with melody and MO idle sound. Melody with Bb2, Db2, Eb3, F3 and Ab3 prefabs working. Also, idle sound is set up for these prefabs. Additionally, the music objects' scripting is moved to scripts on the music objects instead of on the player (from `NoteTrigger` to `MusicObject`). Also, Main Camera is now a prefab.