Added Shield Color change with gradient when triggered by note objects.
Shield changes it's Fresnel Color, when triggered by an object on the Layer "Notes". It takes the Emission Color from those objects. Then it gradually changes the color using a gradient between the current color and the Notes emission color. Modify speed of color change by setting the "Blend Frequency".