Hand tracking issue
There is a bug in the hand tracking section. If the user always uses only one hand, MediaPipe sometimes thinks there are two. It will lead to "Object reference not set to an instance of an object" (see the image). If the user use two hand when the game start, it won't encounter this bug. But if Mario still use one hand it will be a error.