Directory Clean-up + Ufo Handedness Fix
Handedness from Hand-tracking is not always accurate. Sometimes it either returns the wrong value or none at all. Therefore, the temporary fix was to not use it at all.
In the current version of the game, there is no difference between the LeftHand Ufo and the RightHand Ufo. Therefore, we simply always assign the first hand to Left and second hand to Right.